add unit tests for process unit test definition and update location of tests in util functions
